quantizer: add BSD-license header
